docSerialization unit tests
