+++ package.json
@@ -127,7 +127,7 @@
     "@rollup/plugin-terser": "^0.4.4",
     "@size-limit/preset-big-lib": "^11.1.2",
     "@types/child-process-promise": "^2.2.6",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.12",
     "@types/jsdom": "^21.1.6",
     "@types/katex": "^0.16.7",
     "@types/node": "^17.0.31",
@@ -148,7 +148,7 @@
     "eslint-plugin-ft-flow": "^3.0.7",
     "eslint-plugin-header": "^3.1.1",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jest": "^28.4.0",
+    "eslint-plugin-jest": "^28.5.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-lexical": "file:./eslint-plugin",
     "eslint-plugin-no-function-declare-after-return": "^1.1.0",
@@ -165,8 +165,8 @@
     "hermes-parser": "^0.20.1",
     "hermes-transform": "^0.20.1",
     "husky": "^7.0.1",
-    "jest": "^29.4.0",
-    "jest-environment-jsdom": "^29.4.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jsdom": "^24.0.0",
     "lint-staged": "^11.1.0",
     "minimist": "^1.2.5",
